requirements.txt

fastapi
uvicorn
python-jose
passlib[bcrypt]
jinja2
redis
httpx
